auth: EnvVarsCredentialsSource
This new credentials source supports setting credentials using environment
variables whose names follow a particular naming convention.

This is intended as a pragmatic alternative that is easier to use in CI
systems and similar where the de-facto standard is for credentials to be
provided to the execution environment via environment variables, and so
this can help avoid an extra step of collecting those via a credentials
helper.

This follows the same logic as the existing third-party credentials helper
apparentlymart/terraform-credentials-env (where I happen to have been the
"third party") with the intent that this can be as close as possible to a
drop-in replacement for anyone who was already using that, even though
until this point it was just an unofficial solution.