svchost: Specialized error for using a URL as a hostname
There are various reasons why someone might accidentally use a full URL
where we expect a hostname. For example, copying a hostname from the
location bar of a modern web browser tends to silently add the scheme to
the content on the clipboard even though it wasn't disclosed in the UI.

To help diagnose that we'll catch that as a special case and return a
tailored error message for it. The heuristic here is to detech when both
the port number is invalid and when the apparent hostname is "http" and
"https", which should hopefully be loose enough to show this message where
it's helpful (even accommodating slightly-invalid URLs) but minimize
false positives where it might be misleading advice.