disco: Allow concurrent calls to service discovery
Previously the disco.Disco implementation was not concurrency-safe, but
didn't document that as being the case and so some callers tried to use
it concurrently. With the previous implementation that would potentially
corrupt the internal maps if two concurrent calls try to access them at
once.

It's relatively straightforward to use a mutex to avoid corrupting the
maps, so now we'll do that rather than making safe concurrent use the
caller's problem.

This intentionally uses the mutex only to protect the integrity of the
two maps and _does not_ prevent two concurrent network lookups. This means
that it's possible in theory to concurrently run discovery for the same
hostname more than once concurrently if both see the cache unpopulated
when they first check, in which case it's unspecified which of the two
gets commemorated in the cache for future use, but for reasonable server
implementations this shouldn't actually matter in practice.