disco: OverrideHostDiscoveryURL
The disco.Disco type already supports forcing a static service discovery
result for a particular hostname using Disco.ForceHostServices, which
provides one way to override the services on a particular host.

However, using that method requires hard-coding the services and thus uses
of this for non-development purposes effectively subvert the main
motivation of the service discovery system: to have the service information
for a particular hostname encoded in a central location so that it can
be updated everywhere without notifying all clients.

Disco.OverrideHostDiscoveryURL provides a slightly different take on that
that makes a different compromise: instead of specifying a fixed service
discovery result, a caller can instead tell the Disco object to use a
different URL when fetching the service discovery information for a host
over the network. This means that the overridden definition is still
maintained in a central place that can be updated when needed, and in
particular the server that the overridden URL refers to could potentially
produce a servide discovery document that's somehow derived from the
upstream one, such as by overriding the location of just one service while
retaining the others.

However, this should only be done with URLs that the caller trusts, because
this effectively allows redirecting all requests for a particular hostname
to different servers that are under someone else's control, and any
credentials configured for that hostname would still be sent to the new
service endpoints presented in the overridden document.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>